fixed layout and aria issues with MediaTrack
