+++ src/components/media-track/media-track.component.tsx
@@ -1,13 +1,9 @@
 import React, { useCallback } from 'react';
-import { useSelector } from 'react-redux';
 import classNames from 'classnames/bind';
-import _ from 'lodash';
 
 import { useContextMenu, useMediaTrackList } from '../../contexts';
-import { MediaEnums } from '../../enums';
+import { SystemEnums } from '../../enums';
 import { IMediaTrack } from '../../interfaces';
-import { RootState } from '../../reducers';
-import { MediaPlayerService } from '../../services';
 import { DateTimeUtils } from '../../utils';
 
 import { MediaCoverPicture } from '../media-cover-picture/media-cover-picture.component';
@@ -15,6 +11,7 @@ import { MediaTrackInfoComponent } from '../media-track-info/media-track-info.co
 import { MediaPlaybackButton } from '../media-playback-button/media-playback-button.component';
 
 import styles from './media-track.component.css';
+import { useMediaTrackPlayback } from './use-media-track-playback';
 
 const cx = classNames.bind(styles);
 
@@ -26,80 +23,9 @@ export type MediaTrackProps<T> = {
   isPlaying?: boolean,
   disableCover?: boolean,
   disableAlbumLink?: boolean,
+  containerRef?: React.Ref<HTMLDivElement>,
+  containerProps?: React.HTMLProps<HTMLDivElement>,
 };
-
-function useMediaTrackPlayback<T extends IMediaTrack>(props: {
-  mediaTrack: T,
-  mediaTrackPointer?: number,
-  onMediaTrackPlay?: (mediaTrack: T) => void,
-  isPlaying?: boolean, // use the flag to force the playback state, otherwise uses the global playback state
-}) {
-  const {
-    mediaTrack,
-    mediaTrackPointer,
-    onMediaTrackPlay,
-    isPlaying = false,
-  } = props;
-
-  const {
-    mediaPlaybackState,
-    mediaPlaybackCurrentMediaTrack,
-  } = useSelector((state: RootState) => state.mediaPlayer);
-
-  const {
-    mediaTracks,
-    mediaTrackList,
-  } = useMediaTrackList();
-
-  const isTrackPlaying = isPlaying || (mediaPlaybackState === MediaEnums.MediaPlaybackState.Playing
-    && mediaPlaybackCurrentMediaTrack
-    && mediaPlaybackCurrentMediaTrack.tracklist_id === mediaTrackList?.id
-    && mediaPlaybackCurrentMediaTrack.id === mediaTrack.id);
-
-  const play = useCallback(() => {
-    if (onMediaTrackPlay) {
-      onMediaTrackPlay(mediaTrack);
-    } else if (!_.isEmpty(mediaTracks)) {
-      // when playing from a list, media track pointer is required to be provided
-      if (_.isNil(mediaTrackPointer)) {
-        throw new Error('MediaTrackActionButton encountered error while playing track - MediaTrack pointer was not provided');
-      }
-
-      MediaPlayerService.playMediaTrackFromList(mediaTracks, mediaTrackPointer, mediaTrackList);
-    } else {
-      MediaPlayerService.playMediaTrack(mediaTrack);
-    }
-  }, [
-    onMediaTrackPlay,
-    mediaTrack,
-    mediaTrackPointer,
-    mediaTracks,
-    mediaTrackList,
-  ]);
-
-  const pause = useCallback(() => {
-    MediaPlayerService.pauseMediaPlayer();
-  }, []);
-
-  const toggle = useCallback(() => {
-    if (isTrackPlaying) {
-      pause();
-    } else {
-      play();
-    }
-  }, [
-    isTrackPlaying,
-    pause,
-    play,
-  ]);
-
-  return {
-    isTrackPlaying,
-    play,
-    pause,
-    toggle,
-  };
-}
 
 export function MediaTrack<T extends IMediaTrack>(props: MediaTrackProps<T>) {
   const {
@@ -110,6 +36,8 @@ export function MediaTrack<T extends IMediaTrack>(props: MediaTrackProps<T>) {
     isPlaying = false,
     disableCover = false,
     disableAlbumLink = false,
+    containerRef,
+    containerProps = {},
   } = props;
 
   const { showMenu } = useContextMenu();
@@ -145,22 +73,24 @@ export function MediaTrack<T extends IMediaTrack>(props: MediaTrackProps<T>) {
     mediaTrackContextMenuId,
   ]);
 
-  // mediaTrackPointer can be used for providing position for a MediaTrack within a list
-  // this information can be then used for setting accessibility control over our MediaTrack container
-  const mediaTrackAriaProps = !_.isNil(mediaTrackPointer) && {
-    role: 'row',
-    tabIndex: mediaTrackPointer + 1,
-    'aria-rowindex': mediaTrackPointer + 1,
-  };
-
   return (
+    // provide aria props via containerProps, setting anything here causes issues
+    // eslint-disable-next-line jsx-a11y/no-static-element-interactions
     <div
+      {...containerProps}
+      ref={containerRef}
       className={cx('media-track')}
       onContextMenu={handleOnContextMenu}
       onDoubleClick={() => {
         toggle();
       }}
-      {...mediaTrackAriaProps}
+      onKeyDown={(e) => {
+        if (e.key === SystemEnums.KeyboardKeyCodes.Enter
+          || e.key === SystemEnums.KeyboardKeyCodes.Space) {
+          e.preventDefault();
+          toggle();
+        }
+      }}
     >
       <div className={cx('media-track-main-column')}>
         <div className={cx('media-track-section')}>
